Make setupAuth async
