+++ testing/oauth_server.ts
@@ -1,14 +1,12 @@
 import 'cross-fetch/polyfill';
-import {HttpStatusCode} from './constants';
 import type {OAuth2Authentication} from '../types';
-import type {OAuth2ClientCredentialsAuthentication} from '../types';
-import type {OAuth2ClientCredentialsRequestAccessTokenParams} from './auth_types';
 import type {OAuth2RequestAccessTokenParams} from './auth_types';
 import {exec} from 'child_process';
 import express from 'express';
-import {getExpirationDate} from './helpers';
+import {getTokenExpiry} from './oauth_helpers';
 import type * as http from 'http';
 import {print} from './helpers';
+import {requestOAuthAccessToken} from './oauth_helpers';
 import {withQueryParams} from '../helpers/url';
 
 interface AfterTokenOAuthAuthorizationCodeExchangeParams {
@@ -25,72 +23,6 @@ interface AuthorizationCodeTokenCallbackResponse {
 
 export type AfterAuthorizationCodeTokenExchangeCallback =
     (params: AfterTokenOAuthAuthorizationCodeExchangeParams) => void;
-
-interface AfterTokenOAuthClientCredentialsExchangeParams {
-  accessToken: string;
-  expires?: string;
-}
-
-async function requestOAuthAccessToken(
-    params: OAuth2RequestAccessTokenParams
-    | OAuth2ClientCredentialsRequestAccessTokenParams,
-    {tokenUrl, nestedResponseKey, scopeParamName}:
-        {tokenUrl: string; nestedResponseKey?: string, scopeParamName?: string}
-) {
-  const headers = new Headers({
-    'Content-Type': 'application/x-www-form-urlencoded',
-    accept: 'application/json',
-  });
-
-  const formParams = new URLSearchParams();
-  const formParamsWithSecret = new URLSearchParams();
-  for (const [key, value] of Object.entries(params)) {
-    if (value === undefined) {
-      continue;
-    }
-
-    let paramKey = key;
-    if (key === 'scope' && scopeParamName) {
-      paramKey = scopeParamName;
-    }
-    if (paramKey !== 'client_secret') {
-      formParams.append(paramKey, value.toString());
-    }
-
-    formParamsWithSecret.append(paramKey, value.toString());
-  }
-
-  let oauthResponse = await fetch(tokenUrl, {
-    method: 'POST',
-    body: formParamsWithSecret,
-    headers,
-  });
-
-  if (oauthResponse.status === HttpStatusCode.Unauthorized) {
-    // https://datatracker.ietf.org/doc/html/rfc6749#section-3.2.1 doesn't specify how exactly client secret is
-    // passed to the oauth provider. https://datatracker.ietf.org/doc/html/rfc6749#section-2.3 says that client should
-    // NOT has more than one auth methods.
-    //
-    // To workaround with OAuth provider that uses different auth method, we fallback to header auth if body param
-    // auth fails with 401. This is the same behavior in production.
-    headers.append('Authorization', `Basic ${Buffer.from(`${params.client_id}:${params.client_secret}`).toString('base64')}`);
-    oauthResponse = await fetch(tokenUrl, {
-      method: 'POST',
-      body: formParams,
-      headers,
-    });
-  }
-
-  if (!oauthResponse.ok) {
-    throw new Error(`OAuth provider returns error ${oauthResponse.status} ${await oauthResponse.text()}`);
-  }
-
-  const responseBody = await oauthResponse.json();
-  const tokenContainer = nestedResponseKey ? responseBody[nestedResponseKey] : responseBody;
-  const {access_token: accessToken, refresh_token: refreshToken, ...data} = tokenContainer;
-
-  return {accessToken, refreshToken, data};
-}
 
 export function launchOAuthServerFlow({
   clientId,
@@ -155,10 +87,6 @@ export function makeRedirectUrl(port: number): string {
   return `http://localhost:${port}/oauth`;
 }
 
-function _getTokenExpiry(data: {[key: string]: string}) {
-  return  data.expires_in && getExpirationDate(Number(data.expires_in)).toString();
-}
-
 class OAuthServerContainer {
   private readonly _port: number;
   private readonly _afterTokenExchange: AfterAuthorizationCodeTokenExchangeCallback;
@@ -185,7 +113,7 @@ class OAuthServerContainer {
       if (code) {
         const tokenData = await this._tokenCallback(code);
         const {accessToken, refreshToken, data} = tokenData;
-        const expires = _getTokenExpiry(data);
+        const expires = getTokenExpiry(data);
         this._afterTokenExchange({accessToken, refreshToken, expires});
         return res.send('OAuth authentication is complete! You can close this browser tab.');
       }
@@ -200,37 +128,3 @@ class OAuthServerContainer {
   }
 }
 
-export async function performOAuthClientCredentialsServerFlow({
-  clientId,
-  clientSecret,
-  authDef,
-  scopes,
-  afterTokenExchange,
-}: {
-  clientId: string;
-  clientSecret: string;
-  authDef: OAuth2ClientCredentialsAuthentication;
-  scopes?: string[];
-  afterTokenExchange?: (params: AfterTokenOAuthClientCredentialsExchangeParams) => void;
-}): Promise<AfterTokenOAuthClientCredentialsExchangeParams> {
-  const {tokenUrl, nestedResponseKey, scopeParamName, scopeDelimiter} = authDef;
-  // Use the manifest's scopes as a default.
-  const requestedScopes = scopes && scopes.length > 0 ? scopes : authDef.scopes;
-  const scope = requestedScopes ? requestedScopes.join(scopeDelimiter || ' ') : requestedScopes;
-  const params: OAuth2ClientCredentialsRequestAccessTokenParams = {
-    grant_type: 'client_credentials',
-    client_id: clientId,
-    client_secret: clientSecret,
-    scope,
-  };
-
-  const {accessToken, data} = await requestOAuthAccessToken(params, {
-    tokenUrl,
-    nestedResponseKey,
-    scopeParamName,
-  });
-
-  const credentials = {accessToken, expires: _getTokenExpiry(data)};
-  afterTokenExchange?.(credentials);
-  return {accessToken, expires: _getTokenExpiry(data)}
-}
